upadtes

+++ app/Http/Controllers/Backend/RoleController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Module;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 class RoleController extends Controller
 {
@@ -16,6 +17,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.role.index');
         $roles = Role::all();
         return view('backend.role.index', compact(['roles']));
     }
@@ -27,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.role.create');
         $modules = Module::all();
         return view('backend.role.form', compact(['modules']));
     }
@@ -39,6 +42,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.role.create');
         $this->validate($request, [
             'name' => 'required|unique:roles',
             'permissions' => 'required|array',
@@ -73,6 +77,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('app.role.edit');
         $modules = Module::all();
         return view('backend.role.form', compact(['modules','role']));
     }
@@ -86,7 +91,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
-
+        Gate::authorize('app.role.edit');
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name),
@@ -104,6 +109,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('app.role.destroy');
         if ($role->deletable) {
             $role->delete();
             notify()->success("Role Deleted","Success");
